Bridge int<->pointer carrier casts for closure return/arg emit (#287)
Closes the two `-Wint-conversion` warnings in `emit-c` output for the
two-level Signal-Function shape (`sf-let-bind-with-inner-call`). Both are
the same root phenomenon: a value whose Turmeric type is a function/closure
is carried in C as the `int64_t`/`void *` closure carrier, but at one
boundary the emitter omitted the `(int64_t)(intptr_t)` / `(void *)(intptr_t)`
bridge cast.

(1) emit_fns.c -- a `defn` whose inferred result is itself a function type
is lowered to the `int64_t` carrier, but the bare-fn-return guard only fired
for `result_kind == TY_INT`. A closure-returning `defn` (`result_kind ==
TY_FN`) whose C return type is the `int64_t` carrier (e.g. `make-sf`, or the
`__inst_*` closure-return methods) emitted `return <ptr>;` with no cast.
Generalize the guard: compute the function's actual C return type and bridge
a function-typed body whenever the carrier is `int64_t`. This also fixes a
pre-existing latent warning in six instance-closure-return fixtures, whose
old snapshots returned a `void *` box through an `int64_t` return type.

(2) emit_expr.c -- after the two-level-SF fix retypes an already-fat `^fat`
argument to the `:ptr<void>` carrier, the thin local-fn call site cast the
callee's param to `void *` but passed the arg's underlying `int64_t` C
variable uncoerced. Bridge a bare variable reference whose binding lowers to
the int64_t carrier (a `^fat` fn-param) when the formal lowers to a pointer.
Restricted to variable references so a call already returning a pointer
(e.g. `make_hyadder()` -> `void *`) is not gratuitously cast.

Both were benign on LP64 (same width) -- codegen hygiene, not miscompiles.
The fixture still prints 7; full suite green (1493 passed, 0 failed).
Snapshots regenerated for the six fixtures whose latent warning is now fixed.

+++ tests/fixtures/closure-transitive-grandparent-capture/expected.c
@@ -3368,7 +3368,7 @@ static int64_t __fn_898(void * __env_p_901, int64_t x) {
         __t20->x = x;
         __t20->n = __env___env_900->n;
         void *__t21 = __t20;
-        return __t21;
+        return (int64_t)(intptr_t)__t21;
 }
 
 static void * array_hyget(void * arr, int64_t idx) {
@@ -4477,7 +4477,7 @@ static int64_t adder(int64_t n) {
         __t27->__fn = (int64_t)(intptr_t)__fn_898;
         __t27->n = n;
         void *__t28 = __t27;
-        return __t28;
+        return (int64_t)(intptr_t)__t28;
 }
 
 static int64_t call_hyfat(int64_t f, int64_t x) {
